Incorporating additional changes to help with airgapped builds

+++ Vagrantfile
@@ -11,13 +11,41 @@
 # worker_count: number of worker nodes to build (default 3)
 
 require 'securerandom'
+require 'uri'
+
+def get_environment_variable(envvar)
+  ENV[envvar] || ENV[envvar.upcase] || ENV[envvar.downcase]
+end
 
 $generic_install_script = <<-EOH
   #!/bin/sh
-  curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
-  echo 'deb http://apt.kubernetes.io/ kubernetes-xenial main' > /etc/apt/sources.list.d/kubernetes.list
+
+  # Update sources.list
+  if [ -f /vagrant/custom/sources.list ]; then
+    cp /vagrant/custom/sources.list /etc/apt/custom/sources.list
+  fi
+
+  # Update sources.list for kubernetes
+  if [ -f /vagrant/custom/sources.list.kubernetes ]; then
+    cp /vagrant/custom/sources.list.kubernetes /etc/apt/sources.list.d/kubernetes.list
+    KUBEURL=$(cat /vagrant/custom/sources.list.kubernetes | grep '^deb\s' | awk '{ print $2 }' | head -n 1 | sed -e 's/"//g')
+    curl -s "${KUBEURL}/doc/apt-key.gpg" | apt-key add -
+  else
+    curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
+    echo 'deb http://apt.kubernetes.io/ kubernetes-xenial main' > /etc/apt/sources.list.d/kubernetes.list
+  fi
+
+  # Install Docker/Kubernetes
   apt-get update
   apt-get install -y docker.io kubelet kubeadm kubectl kubernetes-cni
+
+  # Override repo prefix for kubelet if needed
+  if [[ "$KUBE_REPO_PREFIX" != "" ]]; then
+    OVERRIDE=/etc/systemd/system/kubelet.service.d/override.conf
+    echo "[Service]\nEnvironment=\"KUBELET_EXTRA_ARGS=--pod-infra-container-image ${KUBE_REPO_PREFIX}/pause-amd64:3.0\"" > $OVERRIDE
+
+    systemctl daemon-reload && sleep 5 && systemctl restart kubelet
+  fi
 EOH
 
 Vagrant.configure("2") do |config|
@@ -25,14 +53,22 @@ Vagrant.configure("2") do |config|
   master_vm = 'master'
   master_vm_ip = '192.168.77.2'
   flannel_network = '10.244.0.0/16'
+  repo_prefix = get_environment_variable('repo_prefix')
+  kubeadm_env = \
+    if repo_prefix
+      {"KUBE_REPO_PREFIX" => URI.join(repo_prefix, 'gcr.io/google_containers').to_s}
+    else
+      {}
+    end
+
   update_ca_certificates_script = ''
 
-  if Vagrant.has_plugin?("vagrant-proxyconf")
-    # copy proxy settings from the Vagrant environment into the VMs
-    http_proxy = ENV['http_proxy'] || ENV['HTTP_PROXY']
-    https_proxy = ENV['https_proxy'] || ENV['HTTPS_PROXY']
-    no_proxy = ENV['no_proxy'] || ENV['NO_PROXY']
+  http_proxy = get_environment_variable('http_proxy')
+  https_proxy = get_environment_variable('https_proxy')
+  no_proxy = get_environment_variable('no_proxy')
 
+  if Vagrant.has_plugin?("vagrant-proxyconf") && (http_proxy || https_proxy || no_proxy)
+    # copy proxy settings from the Vagrant environment into the VMs
     (no_proxy = [no_proxy, master_vm_ip].join(',')) unless no_proxy.nil?
 
     # vagrant-proxyconf will not configure anything if everything is nil
@@ -41,19 +77,20 @@ Vagrant.configure("2") do |config|
     config.proxy.no_proxy = no_proxy
 
     # look for another envvar that points to a directory with CA certs
-    if ENV['cacerts_dir']
-      cacerts_dir = Dir.new(ENV['cacerts_dir'])
+    env_cacerts_dir = get_environment_variable('cacerts_dir')
+    if env_cacerts_dir
+      cacerts_dir = Dir.new(env_cacerts_dir)
       files_in_cacerts_dir = cacerts_dir.entries.select{ |e| not ['.', '..'].include? e }
       files_in_cacerts_dir.each do |f|
         next if File.directory?(File.join(cacerts_dir, f))
         begin
           unless f.end_with? '.crt'
-            fail "All files in #{ENV['cacerts_dir']} must end in .crt due to update-ca-certificates restrictions."
+            fail "All files in #{env_cacerts_dir} must end in .crt due to update-ca-certificates restrictions."
           end
           # read in the certificate and normalize DOS line endings to UNIX
-          cert_raw = File.read(File.join(ENV['cacerts_dir'], f)).gsub(/\r\n/, "\n")
+          cert_raw = File.read(File.join(env_cacerts_dir, f)).gsub(/\r\n/, "\n")
           if cert_raw.scan('-----BEGIN CERTIFICATE-----').length > 1
-            fail "Multiple certificates detected in #{File.join(ENV['cacerts_dir'], f)}, please split them into separate certificates."
+            fail "Multiple certificates detected in #{File.join(env_cacerts_dir, f)}, please split them into separate certificates."
           end
           cert = OpenSSL::X509::Certificate.new(cert_raw) # test that the cert is valid
           dest_cert_path = File.join('/usr/local/share/ca-certificates', f)
@@ -61,13 +98,13 @@ Vagrant.configure("2") do |config|
             echo -ne "#{cert_raw}" > #{dest_cert_path}
           EOH
         rescue OpenSSL::X509::CertificateError
-          fail "Certificate #{File.join(ENV['cacerts_dir'], f)} is not a valid PEM certificate, aborting."
+          fail "Certificate #{File.join(env_cacerts_dir, f)} is not a valid PEM certificate, aborting."
         end # begin/rescue
       end # files_in_cacerts_dir.each
       update_ca_certificates_script << <<-EOH
         update-ca-certificates
       EOH
-    end # if ENV['cacerts_dir']
+    end # if env_cacerts_dir
   end # if Vagrant.has_plugin?
 
   config.vm.box = "ubuntu/xenial64"
@@ -99,6 +136,7 @@ Vagrant.configure("2") do |config|
   end
 
   config.vm.provision 'generic-install', type: 'shell' do |s|
+    s.env = kubeadm_env
     s.inline = $generic_install_script
   end
 
@@ -123,6 +161,7 @@ Vagrant.configure("2") do |config|
     end
 
     c.vm.provision 'kubeadm-init', type: 'shell' do |s|
+      s.env = kubeadm_env
       s.inline = "sudo kubeadm init --apiserver-advertise-address #{master_vm_ip} --pod-network-cidr #{flannel_network} --token #{kubeadm_token}"
     end
 
@@ -137,13 +176,32 @@ Vagrant.configure("2") do |config|
     end
 
     c.vm.provision 'install-flannel', type: 'shell' do |s|
+      s.env = kubeadm_env
       s.privileged = false
       s.inline = <<-EOH
         #!/bin/sh
-        curl -s -O https://raw.githubusercontent.com/coreos/flannel/master/Documentation/kube-flannel-rbac.yml
-        curl -s -O https://raw.githubusercontent.com/coreos/flannel/master/Documentation/kube-flannel.yml
+
+        if [ -f /vagrant/custom/kube-flannel-rbac.yml ]; then
+          cp /vagrant/custom/kube-flannel-rbac.yml $HOME/kube-flannel-rbac.yml
+        else
+          curl -s -O https://raw.githubusercontent.com/coreos/flannel/master/Documentation/kube-flannel-rbac.yml
+        fi
+
+        if [ -f /vagrant/custom/kube-flannel.yml ]; then
+          cp /vagrant/custom/kube-flannel.yml $HOME/kube-flannel.yml
+        else
+          curl -s -O https://raw.githubusercontent.com/coreos/flannel/master/Documentation/kube-flannel.yml
+        fi
+
         export FLANNEL_IFACE=$(ip a | grep #{master_vm_ip} | awk '{ print $NF }')
-        sed -r -i -e "s|command: \\[ \\"/opt/bin/flanneld\\", \\"--ip-masq\\", \\"--kube-subnet-mgr\\" \\]|command: \\[ \\"/opt/bin/flanneld\\", \\"--ip-masq\\", \\"--kube-subnet-mgr\\", \\"--iface\\", \\"${FLANNEL_IFACE}\\" \\]|" kube-flannel.yml
+
+        # substitute in the interface on our VM as the Flannel interface
+        sed -r -i -e "s|command: \\[ \\"/opt/bin/flanneld\\", \\"--ip-masq\\", \\"--kube-subnet-mgr\\" \\]|command: \\[ \\"/opt/bin/flanneld\\", \\"--ip-masq\\", \\"--kube-subnet-mgr\\", \\"--iface\\", \\"${FLANNEL_IFACE}\\" \\]|" $HOME/kube-flannel.yml
+
+        if [[ "$KUBE_REPO_PREFIX" != "" ]]; then
+          sed -i -e "s|quay.io/coreos/flannel:|${KUBE_REPO_PREFIX}/quay.io/coreos/flannel:|g" $HOME/kube-flannel.yml
+        fi
+
         kubectl create -f $HOME/kube-flannel-rbac.yml
         sleep 2
         kubectl create -f $HOME/kube-flannel.yml
@@ -179,6 +237,7 @@ Vagrant.configure("2") do |config|
       end
 
       c.vm.provision 'join-kubernetes-cluster', type: 'shell' do |s|
+        s.env = kubeadm_env
         s.inline = "sudo kubeadm join --token #{kubeadm_token} #{master_vm_ip}:6443"
       end
     end
